+++ MC Engine/IMesh.cpp
@@ -50,7 +50,7 @@ DMesh* ImporterMesh::ImportMesh(aiMesh * buffer, GameObject* object, uint id)
 			mesh->nFaces = newMesh->mNumFaces;
 
 			mesh->nIndex = newMesh->mNumFaces * 3;
-			mesh->Index = new float[mesh->nIndex]; //every face is a triangle.
+			mesh->Index = new uint[mesh->nIndex]; //every face is a triangle.
 
 			for (uint i = 0; i < newMesh->mNumFaces; ++i)
 			{
@@ -593,7 +593,7 @@ DMesh* ImporterMesh::Load(const void* buffer, const char * loadFile, uint id)
 	//--- 
 	cursor += bytes;
 	bytes = sizeof(float) *data->nIndex;
-	data->Index = new float[data->nIndex*3];
+	data->Index = new uint[data->nIndex*3];
 
 	memcpy(data->Index, cursor, bytes);
